backend/requirements.txt
# ============================================
# HoppingRabbit AI - Backend Dependencies V2
# ============================================

# ============================================
# Web Framework
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==12.0

# ============================================
# Task Queue & Message Broker
# ============================================
celery[rabbitmq]==5.4.0
redis==5.2.0
# kombu 由 celery 自动管理
pika==1.3.2
flower==2.0.1

# ============================================
# Database & Storage
# ============================================
supabase==2.10.0

# ============================================
# AI / ML - Core
# ============================================
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper==1.1.0
silero-vad>=5.1

# ============================================
# AI / ML - Audio Separation
# ============================================
demucs==4.0.1

# ============================================
# AI / ML - Speaker Diarization
# ============================================
pyannote.audio==3.1.1

# ============================================
# Video / Audio Processing
# ============================================
ffmpeg-python==0.2.0
soundfile==0.12.1
librosa==0.10.2
scipy>=1.10.0

# ============================================
# Scene Detection
# ============================================
scenedetect[opencv]>=0.6.7

# ============================================
# AI / ML - VAD
# ============================================
webrtcvad==2.0.10

# ============================================
# AI / ML - Speaker Embeddings (Optional)
# ============================================
resemblyzer==0.1.3

# ============================================
# LLM
# ============================================
openai==1.55.0

# ============================================
# HTTP Client
# ============================================
httpx==0.27.2
aiofiles==24.1.0

# ============================================
# Utilities
# ============================================
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-json-logger==2.0.7
structlog==24.4.0
scikit-learn>=1.3.0

# ============================================
# Image Processing
# ============================================
Pillow>=10.0.0
numpy>=1.24.0

